move spherical harmonic recipe module into surface_fitting
